// src/components/library/button/plain/types.ts
export interface PlainButtonPropsI {
    name: string;
    onClick: () => void;
    id?: string;
    bg?: string;
    width?: string;
    height?: string;
    styles?: string;
    margin?: string;
    padding?: string;
    fontSize?: string;
    className?: string;
    disabled?: boolean;
    borderRadius?: string;
}